+++ hooks/src/text_editor.rs
@@ -1,7 +1,6 @@
 use std::{fmt::Display, ops::Range};
 
 use freya_elements::{Code, Key, Modifiers};
-use ropey::iter::Lines;
 pub use ropey::Rope;
 
 /// Holds the position of a cursor in a text
@@ -47,7 +46,7 @@ impl TextCursor {
 /// A text line from a [TextEditor]
 #[derive(Clone)]
 pub struct Line<'a> {
-    text: &'a str,
+    pub text: &'a str,
 }
 
 impl Line<'_> {
@@ -298,92 +297,5 @@ pub trait TextEditor: Sized + Clone + Display {
         }
 
         event
-    }
-}
-
-#[derive(Clone)]
-pub struct UseEditableText {
-    rope: Rope,
-    cursor: TextCursor,
-}
-
-impl Display for UseEditableText {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str(&self.rope.to_string())
-    }
-}
-
-impl TextEditor for UseEditableText {
-    type LinesIterator<'a> = LinesIterator<'a>;
-
-    fn lines(&self) -> Self::LinesIterator<'_> {
-        let lines = self.rope.lines();
-        LinesIterator { lines }
-    }
-
-    fn insert_char(&mut self, char: char, char_idx: usize) {
-        self.rope.insert_char(char_idx, char);
-    }
-
-    fn insert(&mut self, text: &str, char_idx: usize) {
-        self.rope.insert(char_idx, text);
-    }
-
-    fn remove(&mut self, range: Range<usize>) {
-        self.rope.remove(range)
-    }
-
-    fn char_to_line(&self, char_idx: usize) -> usize {
-        self.rope.char_to_line(char_idx)
-    }
-
-    fn line_to_char(&self, line_idx: usize) -> usize {
-        self.rope.line_to_char(line_idx)
-    }
-
-    fn line(&self, line_idx: usize) -> Option<Line<'_>> {
-        let line = self.rope.get_line(line_idx);
-
-        line.map(|line| Line {
-            text: line.as_str().unwrap_or(""),
-        })
-    }
-
-    fn len_lines<'a>(&self) -> usize {
-        self.rope.len_lines()
-    }
-
-    fn cursor(&self) -> &TextCursor {
-        &self.cursor
-    }
-
-    fn cursor_mut(&mut self) -> &mut TextCursor {
-        &mut self.cursor
-    }
-}
-
-/// Iterator over text lines.
-pub struct LinesIterator<'a> {
-    lines: Lines<'a>,
-}
-
-impl<'a> Iterator for LinesIterator<'a> {
-    type Item = Line<'a>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        let line = self.lines.next();
-
-        line.map(|line| Line {
-            text: line.as_str().unwrap_or(""),
-        })
-    }
-}
-
-impl From<&str> for UseEditableText {
-    fn from(value: &str) -> Self {
-        Self {
-            rope: Rope::from_str(value),
-            cursor: TextCursor::new(0, 0),
-        }
     }
 }
